SentinelForge/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-socketio==5.10.0
pydantic==2.5.2
numpy==1.26.2
torch==2.1.2
redis==5.0.1
PyJWT==2.8.0
bcrypt==3.2.2
passlib[bcrypt]==1.7.4
transformers==4.35.2
python-multipart==0.0.6

psutil==5.9.8
pywin32==306; sys_platform == 'win32'
wmi==1.5.1; sys_platform == 'win32'


scikit-learn==1.3.2

